% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.rweka-package}
\alias{mlr3learners.rweka}
\alias{mlr3learners.rweka-package}
\title{mlr3learners.rweka: JRip and J48 Learner for mlr3}
\description{
JRip and J48 Learner for mlr3.
}
\author{
\strong{Maintainer}: Henri Funk \email{henrifnk@gmail.com}

}
